lots of Gladiator changes

+++ RandomRPG/RandomRPG/Model/Interfaces/IWeapon.cs
@@ -2,7 +2,7 @@ namespace RandomRPG.Model.Interfaces
 {
     public interface IWeapon : IEquippable
     {
-        double DamageOutput();
+        int DamageOutput();
         int MinDamage { get; set; }
         int MaxDamage { get; set; }
     }
